requirements.txt
fastapi
uvicorn
python-multipart
requests
python-docx
Pillow
